fix: show swap kline as dialog on mobile (OK-55532)
